pyproject.toml
[project]
name = "nijigasaki-whisper"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "authlib>=1.5.2",
    "ffmpeg-python>=0.2.0",
    "filetype>=1.2.0",
    "openai>=1.78.1",
    "streamlit>=1.45.1",
]
